requirements.txt
pyodbc>=4.0.34
pandas>=1.3.5
numpy>=1.21.6
pyyaml>=6.0
tqdm>=4.64.1